refactor(kamn-node): decompose observability endpoint modules (#5180)
